update migrations to use bigint

+++ src/main/resources/stockdinkan/jdbc/V0010__CreateDayValues.sql
@@ -1,12 +1,12 @@
 CREATE TABLE DayValues(
   id SERIAL PRIMARY KEY,
   symbol VARCHAR(30) NOT NULL,
-  stime INTEGER NOT NULL,
+  stime BIGINT NOT NULL,
   sopen FLOAT NOT NULL,
   sclose FLOAT NOT NULL,
   low FLOAT NOT NULL,
   high FLOAT NOT NULL,
-  volume INTEGER NOT NULL,
+  volume BIGINT NOT NULL,
   trade_count INTEGER NOT NULL,
   vwap FLOAT NOT NULL
 );
