+++ src/main/scala/models/DayData.scala
@@ -1,5 +1,6 @@
 package models
 
-case class DayData(id: Int, symbol: String, stime: Long, sopen: Float, sclose: Float, low: Float, high: Float, volume: Int, trade_count: Int, vwap: Float)
+case class DayData(id: Int, symbol: String, stime: Long, sopen: Double, sclose: Double, low: Double, high: Double, volume: Long, trade_count: Int, vwap: Double)
 
 case class DayDataExtras(time: Long, symbol: String)
+
